requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-groq
python-dotenv
pydantic
ipykernel